# docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: InteriorAI API
  version: 1.1.0
  description: >-
    權威契約檔（ADR-015）；同鍵不同內容的 Idempotency-Key 回 422 IDEMPOTENCY_KEY_REUSED（ADR-014）。
    每個 operation 以 `x-phase` 標註交付階段：API 必須實作 x-phase ≤ 目前階段的所有 operation，
    且不得有契約外的路由（契約測試 services/api/test/contract 驗證）。請求本文由 API 依本檔以 JSON Schema 驗證。
servers: [{ url: /v1 }]
security: [{ bearerAuth: [] }]
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    OrgId: { name: X-Org-Id, in: header, required: false, description: 切換作用中的組織（須為成員）, schema: { type: string, format: uuid } }
  schemas:
    ErrorCode:
      type: string
      enum: [AUTH_REQUIRED, FORBIDDEN, NOT_FOUND, VALIDATION_FAILED, VERSION_CONFLICT, INSUFFICIENT_CREDITS,
             RATE_LIMITED, UPLOAD_REJECTED, MODERATION_BLOCKED, JOB_FAILED, PROVIDER_UNAVAILABLE,
             SCHEMA_UNSUPPORTED, SCENE_LIMIT_EXCEEDED, IDEMPOTENCY_KEY_REUSED, SCALE_UNKNOWN,
             CONFLICT, CREDITS_FROZEN, INTERNAL]
    Error:
      type: object
      required: [code, message, requestId]
      properties: { code: { $ref: '#/components/schemas/ErrorCode' }, message: { type: string }, details: {}, requestId: { type: string } }
...
    Project:
      type: object
      required: [id, name, updatedAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        thumbnailUrl: { type: [string, 'null'] }
        currentVersionId: { type: [string, 'null'], format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectCreate:
      type: object
      additionalProperties: false
      required: [name]
      properties: { name: { type: string, minLength: 1, maxLength: 120 } }
    ProjectPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties: { name: { type: string, minLength: 1, maxLength: 120 } }
    ProjectPage:
      type: object
      required: [items, nextCursor]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Project' } }, nextCursor: { type: [string, 'null'] } }
    Version:
      type: object
      required: [id, projectId, schemaVersion, contentHash, createdAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        parentVersionId: { type: [string, 'null'], format: uuid }
        schemaVersion: { type: string }
        contentHash: { type: string, pattern: '^[0-9a-f]{64}$' }
        note: { type: [string, 'null'] }
        createdBy: { type: [string, 'null'], format: uuid }
        createdAt: { type: string, format: date-time }
    VersionWithScene:
      allOf:
        - $ref: '#/components/schemas/Version'
        - type: object
          required: [scene]
          properties: { scene: { $ref: '#/components/schemas/Scene' } }
    VersionCreate:
      type: object
      additionalProperties: false
      required: [baseVersionId, scene]
      properties:
        baseVersionId: { type: [string, 'null'], format: uuid }
        scene: { $ref: '#/components/schemas/Scene' }
        note: { type: string, maxLength: 200 }
...
    Scene: { $ref: 'scene.schema.json' }
...
  responses:
    Err: { description: Error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
paths:
...
  /projects:
    get:
      x-phase: P3
      summary: 專案列表
      parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/OrgId' }]
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectPage' } } } } }
    post:
      x-phase: P3
      summary: 建立專案
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectCreate' } } } }
      responses:
        '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '403': { $ref: '#/components/responses/Err' }
  /projects/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      x-phase: P3
      summary: 取得專案
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '404': { $ref: '#/components/responses/Err' }
    patch:
      x-phase: P3
      summary: 更新專案中繼資料
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectPatch' } } } }
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } } }
    delete: { x-phase: P3, summary: 刪除專案（軟刪除）, responses: { '204': { description: deleted } } }
  /projects/{id}/versions:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      x-phase: P3
      summary: 版本列表（新到舊）
      responses: { '200': { description: ok, content: { application/json: { schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Version' } } } } } } } }
    post:
      x-phase: P3
      summary: 儲存新版本（樂觀鎖；內容與目前版本相同時回 200 與既有版本）
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/VersionCreate' } } } }
      responses:
        '200': { description: 內容未變，回目前版本, content: { application/json: { schema: { $ref: '#/components/schemas/Version' } } } }
        '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/Version' } } } }
        '409': { $ref: '#/components/responses/Err' }
        '422': { $ref: '#/components/responses/Err' }
  /versions/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      x-phase: P3
      summary: 取得版本（含 Scene）
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/VersionWithScene' } } } } }
  /projects/{id}/restore/{versionId}:
    parameters: [{ $ref: '#/components/parameters/Id' }, { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }]
    post:
      x-phase: P3
      summary: 還原版本（把目前版本指回該版本；版本本身不可變）
      responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } } }
...